# frontend/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Synapse API",
    "version": "1.0.0",
    "description": "Shared context layer for AI tools — persistent memory across sessions."
  },
  "servers": [{ "url": "https://api.synapsesync.app" }],
  "security": [{ "BearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from account page or CLI login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        },
        "required": ["error", "code"]
      },
...
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create account",
        "security": [],
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "email": { "type": "string" },
                    "api_key": { "type": "string" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login and get API key",
        "security": [],
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" },
                  "label": { "type": "string", "default": "cli" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "email": { "type": "string" },
                    "api_key": { "type": "string" },
                    "label": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  }
}